refactors blockparser: moves cleaning stuff into its own class
