+++ Quagga/__init__.py
@@ -1,6 +1,6 @@
 
 from Quagga.__main__ import Quagga
-from Quagga.Utils.EmailDirectoryReader import EmailDirectoryReader, TempQuaggaReader
-from Quagga.Utils.EmailListReader import ListReaderRawEmailTexts, ListReaderExtractedBodies
+from Quagga.Utils.Reader.EmailDirectoryReader import EmailDirectoryReader, TempQuaggaReader
+from Quagga.Utils.Reader.EmailListReader import ListReaderRawEmailTexts, ListReaderExtractedBodies
 
 from Quagga.Utils.ModelBuilder import ModelBuilder
